% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_format}
\alias{check_format}
\title{Detect format of the file}
\usage{
check_format(fname)
}
\arguments{
\item{fname}{the file to detect the format of}
}
\description{
Detect format of the file
}
