% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_sequence}
\alias{extract_sequence}
\title{Extract the sequences for a given region}
\usage{
extract_sequence(bam, index, reg)
}
\arguments{
\item{bam}{the cram/bam/sam file}

\item{index}{the index of the cram/bam/sam file}

\item{reg}{the region of interest, typically in format of chr1:start-begin}
}
\value{
a character vector with the sequences in the given region
}
\description{
Extract the sequences for a given region
}
\examples{
\dontrun{count_kmer(bam, index, "chr1:10001-100050")}
}
