% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_content}
\alias{gc_content}
\title{Calculate the GC content for a region}
\usage{
gc_content(bam, index, reg)
}
\arguments{
\item{bam}{the cram/bam/sam file}

\item{index}{the index of the cram/bam/sam file}

\item{reg}{the region of interest, typically in format of chr1:start-begin}
}
\description{
Calculate the GC content for a region
}
