\name{htslibr-package}
\alias{htslibr-package}
\alias{htslibr}
\docType{package}
\title{\packageTitle{htslibr}}
\description{\packageDescription{htslibr}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{htslibr}
  \packageIndices{htslibr}

}
\author{
  \packageAuthor{htslibr}
  
  Maintainer: \packageMaintainer{htslibr}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
