% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_kmer}
\alias{count_kmer}
\title{count the number of times a kmer is present in a region}
\usage{
count_kmer(bam, index, reg, kmer)
}
\arguments{
\item{bam}{the cram/bam/sam file}

\item{index}{the index of the cram/bam/sam file}

\item{reg}{the region of interest, typically in format of chr1:start-begin}

\item{kmer}{the substring to search for in the reads}
}
\value{
a dataframe with the sequnce reads and counts of the given kmer per read (i.e. two columns)
}
\description{
count the number of times a kmer is present in a region
}
\examples{
\dontrun{count_kmer(bam, index, "chr1:10001-100050", "TTACGG")}
}
