% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{depth}
\alias{depth}
\title{Estimate the depth for each position in a given region}
\usage{
depth(bam, index, reg, flank_bp)
}
\arguments{
\item{bam}{the cram/bam/sam file}

\item{index}{the index of the cram/bam/sam file}

\item{reg}{the region of interest, typically in format of chr1:start-begin}
}
\description{
Estimate the depth for each position in a given region
}
