% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_info}
\alias{extract_info}
\title{extract values from the INFO field}
\usage{
extract_info(vcf, index, reg, tag)
}
\arguments{
\item{vcf}{the VCF/BCF file path}

\item{index}{the CSI/TBI index file path}

\item{reg}{a region query of the form: chr:start-end}

\item{tag}{the field in the INFO field to extract. Only accepts one string value at this time. Only can extract numeric fields at the moment.}
}
\value{
a dataframe with the chrom, pos, and value of the given INFO field
}
\description{
Use this function to extract the INFO field values for a single INFO field in a give
region based query.
}
\examples{
\dontrun{extract_info(vcf, index, "1:10001-100500", "AC")}
}
